% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_lib.R
\name{AccumulateInfluenceDf}
\alias{AccumulateInfluenceDf}
\title{Calculate the estimated cumulative effects for a sorted gradient dataframe.}
\usage{
AccumulateInfluenceDf(
  df,
  influence_cols,
  base_vals,
  change_colnames,
  estimate_colnames
)
}
\description{
Calculate the estimated cumulative effects for a sorted gradient dataframe.
}
