% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_iv_grads_lib.R
\name{AppendIntervalColumns}
\alias{AppendIntervalColumns}
\title{Compute the gradients of the upper and lower bounds of confidence intervals
from the gradients of \code{betahat} and \code{se} with respect to the weights.}
\usage{
AppendIntervalColumns(grad_df, sig_num_ses)
}
\arguments{
\item{grad_df}{The output of \code{\link[=ComputeModelInfluence]{ComputeModelInfluence()}}}

\item{sig_num_ses}{How wide the confidence interval is, as a number of standard errors.}
}
\value{
\code{grad_df}, augmented with the interval derivatives and attributes.
}
\description{
Compute the gradients of the upper and lower bounds of confidence intervals
from the gradients of \code{betahat} and \code{se} with respect to the weights.
}
