% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_targets_lib.R
\name{AppendTargetRegressorInfluence}
\alias{AppendTargetRegressorInfluence}
\title{Compute the influence scores for a particular parameter.}
\usage{
AppendTargetRegressorInfluence(
  model_grads,
  target_parameter,
  sig_num_ses = qnorm(0.975)
)
}
\arguments{
\item{model_grads}{A model gradient object (e.g. as produced by \link{ComputeModelInfluence})}

\item{target_parameter}{The string naming a regressor (must be an
entry in \link{model_grads$parameter_names}).}

\item{sig_num_ses}{How wide the confidence interval is, as a number of standard errors.}
}
\value{
The original \code{model_grads}, with an entry
\code{model_grads$param_infl_list[[target_parameter]]} containing a
parameter influence object.
}
\description{
Compute the influence scores for a particular parameter.
}
