% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_targets_lib.R
\name{as.data.frame.qoi_signal}
\alias{as.data.frame.qoi_signal}
\title{Produce a tidy dataframe summarizing a signal.}
\usage{
\method{as.data.frame}{qoi_signal}(signal)
}
\arguments{
\item{signal}{A signal object (e.g. one of the elements of the list produced by \link{get_inference_signals_for_parameter}).}
}
\description{
Produce a tidy dataframe summarizing a signal.
}
