% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_targets_lib.R
\name{as.data.frame.QOISignal}
\alias{as.data.frame.QOISignal}
\title{Produce a tidy dataframe summarizing a signal.}
\usage{
\method{as.data.frame}{QOISignal}(signal)
}
\arguments{
\item{signal}{A signal object (e.g. one of the elements of the list produced by \link{GetInferenceSignals}).}
}
\description{
Produce a tidy dataframe summarizing a signal.
}
