% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_grads_lib.R
\name{CheckLogitDiagnostics}
\alias{CheckLogitDiagnostics}
\title{Check diagnostics for a logistic regression fit.}
\usage{
CheckLogitDiagnostics(glm_res)
}
\arguments{
\item{glm_res}{The result of glm() with family=binomial.}
}
\value{
A list with fitted probabilities, Hessian, reciprocal condition
number, and minimum of p and 1-p.
}
\description{
Check diagnostics for a logistic regression fit.
}
