% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_iv_grads_torch_lib.R
\name{compute_iv_regression_results}
\alias{compute_iv_regression_results}
\title{Run an IV regression using zaminfluence code.  This should be identical
to ordinary regression.}
\usage{
compute_iv_regression_results(iv_res, weights = NULL, se_group = NULL)
}
\arguments{
\item{iv_res}{The iv regression result, i.e, the output of \code{\link[ivreg:ivreg]{ivreg::ivreg()}}.}

\item{weights}{Optional.  A vector of weights.  If unset, use the original weights.}

\item{se_group}{Optional. The standard error grouping variable.}
}
\value{
A list containing the regression coefficients and standard errors.
}
\description{
Run an IV regression using zaminfluence code.  This should be identical
to ordinary regression.
}
