% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_iv_grads_lib.R
\name{ComputeIVRegressionErrorCovariance}
\alias{ComputeIVRegressionErrorCovariance}
\title{Compute the standard error matrix for an IV regression.}
\usage{
ComputeIVRegressionErrorCovariance(iv_res, se_group = NULL)
}
\arguments{
\item{iv_res}{The output of an IV regression computed with AER::ivreg.}

\item{se_group}{Optional, a vector of integers defining a standard error grouping.}
}
\value{
The standard error matrix.
}
\description{
Compute the standard error matrix for an IV regression.
}
