% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_lib.R
\name{ComputeIVRegressionErrorCovariancePython}
\alias{ComputeIVRegressionErrorCovariancePython}
\title{Deprecated --- use the functions in \code{ols_iv_grads_lib.R} instead.}
\usage{
ComputeIVRegressionErrorCovariancePython(iv_res, se_group = NULL)
}
\description{
Deprecated --- use the functions in \code{ols_iv_grads_lib.R} instead.
}
