% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_iv_grads_torch_lib.R
\name{ComputeIVRegressionInfluence}
\alias{ComputeIVRegressionInfluence}
\title{Compute all influence scores for an IV regression.}
\usage{
ComputeIVRegressionInfluence(iv_res, se_group = NULL, keep_pars = NULL)
}
\arguments{
\item{iv_res}{The iv regression result, i.e, the output of \code{\link[AER:ivreg]{ivreg()}}.}

\item{se_group}{Optional. The standard error grouping variable.}

\item{keep_pars}{Optional. The names of parameters for which to compute derivatives.}
}
\value{
A list containing the sensitivity of the coefficients and standard errors to data re-weighting.  The list entries are
\describe{
\item{model_fit}{The original fit as returned by the fitting function}
\item{num_obs}{The number of observations given weights}
\item{parameter_names}{The names of the regressors as strings}
\item{grad_fun}{The function used to compute the derivatives}
\item{param}{The estimated regression coefficient at the original weights}
\item{se}{The standard errors at the original weights}
\item{weights}{The original weights}
\item{param_grad}{The gradient of param with respect to the weights}
\item{se_grad}{The gradient of se with respect to the weights}
}
}
\description{
Compute all influence scores for an IV regression.
}
