% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_lib.R
\name{ComputeIVRegressionInfluencePython}
\alias{ComputeIVRegressionInfluencePython}
\title{Deprecated --- use the functions in \code{ols_iv_grads_lib.R} instead.
Compute all influence scores for an IV regression.}
\usage{
ComputeIVRegressionInfluencePython(iv_res, se_group = NULL)

ComputeIVRegressionInfluencePython(iv_res, se_group = NULL)
}
\arguments{
\item{iv_res}{The output of an IV regression computed with AER::ivreg.}

\item{se_group}{Optional, a vector of integers defining a standard error grouping.}

\item{df}{A dataframe with regression data.}
}
\value{
A list containing the regression and influence result..
}
\description{
Deprecated --- use the functions in \code{ols_iv_grads_lib.R} instead.
Compute all influence scores for an IV regression.

Deprecated --- use the functions in \code{ols_iv_grads_lib.R} instead.
}
