% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_iv_grads_lib.R
\name{ComputeIVRegressionResults}
\alias{ComputeIVRegressionResults}
\title{Run an IV regression using zaminfluence code.  This should be identical
to ordinary regression.}
\usage{
ComputeIVRegressionResults(iv_res, weights = NULL, se_group = NULL)
}
\arguments{
\item{weights}{Optional, a vector of weights.  If unset, use the original
regression weights.}

\item{se_group}{Optional, a vector of integers defining a standard error grouping.}

\item{lm_result}{The output of a call to lm.}
}
\value{
A list containing the regression coefficients and standard errors.
}
\description{
Run an IV regression using zaminfluence code.  This should be identical
to ordinary regression.
}
