% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_iv_grads_torch_lib.R
\name{ComputeModelInfluence}
\alias{ComputeModelInfluence}
\title{Compute the influence functions for all regressors given a model fit.}
\usage{
ComputeModelInfluence(fit_object, keep_pars = NULL, se_group = NULL)
}
\arguments{
\item{se_group}{Optional. The standard error grouping variable.}

\item{model_fit}{The fit from \code{\link[=lm]{lm()}} or \code{\link[AER:ivreg]{AER::ivreg()}}.}
}
\value{
A model gradient object (e.g. as produced by \link{ComputeModelInfluence})
}
\description{
Compute the influence functions for all regressors given a model fit.
}
