% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_lib.R
\name{ComputeRegressionInfluence}
\alias{ComputeRegressionInfluence}
\title{Compute all influence scores for a regression.}
\usage{
ComputeRegressionInfluence(lm_result, se_group = NULL)
}
\arguments{
\item{lm_result}{The output of a call to lm.}

\item{se_group}{Optional, a vector of integers defining a standard error grouping.}
}
\value{
A list containing the regression and influence result.
}
\description{
Compute all influence scores for a regression.
}
