% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_lib.R
\name{ComputeRegressionMomentSensitivity}
\alias{ComputeRegressionMomentSensitivity}
\title{Compute the sensitivity of a regression to the moment condition on the
residuals.  That is, compute the gradients with respect to `offset`,
where the regression is fit with the moment condition Cov(x, resid) = offset.}
\usage{
ComputeRegressionMomentSensitivity(lm_result, se_group = NULL)
}
\arguments{
\item{lm_result}{The output of a call to lm.}

\item{se_group}{Optional, a vector of integers defining a standard error grouping.}
}
\value{
A list containing the regression and influence result.
}
\description{
Compute the sensitivity of a regression to the moment condition on the
residuals.  That is, compute the gradients with respect to `offset`,
where the regression is fit with the moment condition Cov(x, resid) = offset.
}
