% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_iv_grads_lib.R
\name{ComputeRegressionResults}
\alias{ComputeRegressionResults}
\title{Run a regression using zaminfluence code.  This should be identical
to ordinary regression.}
\usage{
ComputeRegressionResults(lm_result, weights = NULL, se_group = NULL)
}
\arguments{
\item{lm_result}{The output of a call to lm.}

\item{weights}{Optional, a vector of weights.  If unset, use the original regression weights.}

\item{se_group}{Optional, a vector of integers defining a standard error grouping.}
}
\value{
A list containing the regression coefficients and standard errors.
}
\description{
Run a regression using zaminfluence code.  This should be identical
to ordinary regression.
}
