% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_iv_grads_lib.R
\name{CopyGradAttributes}
\alias{CopyGradAttributes}
\title{Copy the attributes needed for sorting.}
\usage{
CopyGradAttributes(
  new_df,
  grad_df,
  attrs = c("n_obs", "obs_per_row_col", "base_vals", "target_regressor",
    "target_index", "sig_num_ses", "data_row_cols", "n_grad_rows")
)
}
\arguments{
\item{new_df}{The dataframe to receive the attributes}

\item{grad_df}{The gradient dataframe with attributes for a particular coefficient, e.g., as returned by \code{\link[=GetTargetRegressorGrads]{GetTargetRegressorGrads()}}.}
}
\value{
\code{new_df}, but with the needed attributes of \code{grad_df} set.
}
\description{
Sometimes if you copy or modify the gradient dataframe returned by
\code{\link[=GetTargetRegressorGrads]{GetTargetRegressorGrads()}}, e.g. by merging with another data source,
the attributes required for running \code{\link[=SortAndAccumulate]{SortAndAccumulate()}} are lost.
This function copies the required attributes over from the original
dataframe.
}
