% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_lib.R
\name{get_amis}
\alias{get_amis}
\title{Compute the approximate maximally-influential set (AMIS).}
\usage{
get_amis(qoi, sign, n_drop)
}
\arguments{
\item{qoi}{\verb{A quantity of interest object (e.g. as produced by [qoi_influence]).}}

\item{sign}{Either \code{"pos"} or \code{"neg"}: whether to return the points with
the most positive or the most negative influence scores.}

\item{n_drop}{The number of points to drop (we will round up).}
}
\value{
The indices to drop (in the order of the original data)
}
\description{
Compute the approximate maximally-influential set (AMIS).
}
