% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_targets_lib.R
\name{get_base_values}
\alias{get_base_values}
\title{Extract the base values of each QOI from a parameter influence object.}
\usage{
get_base_values(param_infl)
}
\arguments{
\item{param_infl}{A parameter influence object (e.g. as produced by \link{append_target_regressor_influence}).}
}
\value{
A named numeric vector with entries \code{param}, \code{se}, \code{param_mzse},
\code{param_pzse} giving the QOIs at the original fit.
}
\description{
Extract the base values of each QOI from a parameter influence object.
}
