% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_targets_lib.R
\name{get_inference_signals}
\alias{get_inference_signals}
\title{Compute the signals for changes to sign, significance, and both.}
\usage{
get_inference_signals(model_grads)
}
\arguments{
\item{model_grads}{A model gradient object (e.g. as produced by \link{compute_model_influence})}
}
\value{
A list lists of of signals, one for each parameter in
named model_grads$param_infls.  See the output of
\code{get_inference_signals_for_parameter}.
}
\description{
Compute the signals for changes to sign, significance, and both.
}
