% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_targets_lib.R
\name{get_inference_signals_for_parameter}
\alias{get_inference_signals_for_parameter}
\title{Compute the signals for changes to sign, significance, and both.}
\usage{
get_inference_signals_for_parameter(param_infl)
}
\arguments{
\item{param_infl}{A parameter influence object (e.g. as produced by \link{append_target_regressor_influence}).}
}
\value{
A list of signals, named "sign", "sig", and "both".  Each
entry is a \code{qoi_signal} object.
}
\description{
Compute the signals for changes to sign, significance, and both.
}
