% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_lib.R
\name{GetAMIP}
\alias{GetAMIP}
\title{Compute the approximate maximum influence perturbation (AMIP).}
\usage{
GetAMIP(qoi, n_drop = NULL, prop_drop = NULL)
}
\arguments{
\item{qoi}{\verb{A quantity of interest object (e.g. as produced by [ProcessInfluenceVector]).}}

\item{n_drop}{The number of points to drop (we will round up).
Exactly one of \code{n_drop} and \code{prop_drop} must be specified.}

\item{prop_drop}{The proportion of points to drop (we will round up).
Exactly one of \code{n_drop} and \code{prop_drop} must be specified.}
}
\value{
The approximate largest change that can be produced by dropping
the specified number of points.
}
\description{
Compute the approximate maximum influence perturbation (AMIP).
}
