% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_lib.R
\name{GetAPIP}
\alias{GetAPIP}
\title{Compute the approximate perturbation-inducing proportion (APIP).}
\usage{
GetAPIP(qoi, signal)
}
\arguments{
\item{qoi}{\verb{A quantity of interest object (e.g. as produced by [ProcessInfluenceVector]).}}

\item{signal}{A signal object (e.g. one of the elements of the list produced by \link{GetInferenceSignals}).}
}
\value{
A list containing
\code{n}: The number of points to drop
\code{prop}: The proportion of points to drop
\code{inds}: The indices to drop (in the order of the original data)
}
\description{
Compute the approximate perturbation-inducing proportion (APIP).
}
