% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_lib.R
\name{GetAPIPForQOI}
\alias{GetAPIPForQOI}
\title{Compute the approximate perturbation-inducing proportion (APIP).}
\usage{
GetAPIPForQOI(qoi, signal)
}
\arguments{
\item{qoi}{\verb{A quantity of interest object (e.g. as produced by [QOIInfluence]).}}

\item{signal}{The desired difference.}
}
\value{
A list containing
\code{n}: The number of points to drop
\code{prop}: The proportion of points to drop
\code{inds}: The indices to drop (in the order of the original data)
}
\description{
Compute the approximate perturbation-inducing proportion (APIP).
}
