% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_sensitivity_lib.R
\name{GetGammaAlpha}
\alias{GetGammaAlpha}
\title{Return the $Gamma_alpha$ quantity.}
\usage{
GetGammaAlpha(influence_df, alpha, alpha_col, gamma_alpha_col)
}
\arguments{
\item{influence_df}{A single influence dataframe, e.g., a single member of the list returned by \code{\link[=SortAndAccumulate]{SortAndAccumulate()}}.}

\item{alpha}{The alpha value to target corresponding to \code{alpha_colname}.}
}
\value{
$Gamma_alpha$, the shape coefficient.
}
\description{
Return the $Gamma_alpha$ quantity.
}
