% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_targets_lib.R
\name{GetInferenceSignals}
\alias{GetInferenceSignals}
\title{Compute the signals for changes to sign, significance, and both.}
\usage{
GetInferenceSignals(model_grads)
}
\arguments{
\item{model_grads}{A model gradient object (e.g. as produced by \link{ComputeModelInfluence})}
}
\value{
A list lists of of signals, one for each parameter in
named model_grads$param_infls.  See the output of
\code{GetInferenceSignalsForParameter}.
}
\description{
Compute the signals for changes to sign, significance, and both.
}
