% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_targets_lib.R
\name{GetInferenceSignalsForParameter}
\alias{GetInferenceSignalsForParameter}
\title{Compute the signals for changes to sign, significance, and both.}
\usage{
GetInferenceSignalsForParameter(param_infl)
}
\arguments{
\item{param_infl}{A parameter influence object (e.g. as produced by \link{AppendTargetRegressorInfluence}).}
}
\value{
A list of signals, named "sign", "sig", and "both".  Each
entry is a \code{QOISignal} object.
}
\description{
Compute the signals for changes to sign, significance, and both.
}
