% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_sensitivity_lib.R
\name{GetInfluenceScale}
\alias{GetInfluenceScale}
\title{Calcualate the scale (L2 norm) of an influence vector.}
\usage{
GetInfluenceScale(influence_vec)
}
\arguments{
\item{influence_vec:}{A vector of influence scores.}
}
\value{
$sigma_xi$, the "scale" of the influence function.
}
\description{
Calcualate the scale (L2 norm) of an influence vector.
}
