% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_lib.R
\name{GetModelFitInferenceDataframe}
\alias{GetModelFitInferenceDataframe}
\title{Summarize the values of each QOI for each parameter for a given model_fit.}
\usage{
GetModelFitInferenceDataframe(model_fit, param_infls)
}
\arguments{
\item{model_fit}{The fit from \code{\link[=lm]{lm()}} or \code{\link[AER:ivreg]{AER::ivreg()}}.}

\item{param_infls}{A list of ParameterInferenceInfluence objects.}
}
\value{
A dataframe summarizing the values of all quantities of interest
in \code{param_infls} for \code{model_fit}.
}
\description{
Summarize the values of each QOI for each parameter for a given model_fit.
}
