% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_sensitivity_lib.R
\name{GetRegressionTargetChange}
\alias{GetRegressionTargetChange}
\title{Estimate the number and proportion of datapoints needed to effect various
changes.}
\usage{
GetRegressionTargetChange(influence_dfs, alpha_col)
}
\arguments{
\item{influence_dfs}{A list of sorted influence dataframes as returned, e.g., by \code{\link[=SortAndAccumulate]{SortAndAccumulate()}}.  The list is expected to have results sorted for both negative and positive changes in both sign and signifiance.}

\item{alpha_col}{A string with the name of the target alpha column (typically \code{prop_removed} or \code{num_removed})}
}
\value{
A dataframe summarizing the estimated number of points needed
to effect changes in sign, significance, or both sign and significance,
or \code{NA} if the linear approximation cannot produce such a change.
}
\description{
Estimate the number and proportion of datapoints needed to effect various
changes.
}
