% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_lib.R
\name{GetRerunBaseValues}
\alias{GetRerunBaseValues}
\title{Compute the "base values" (beta and the CI) for a parameter from a rerun.}
\usage{
GetRerunBaseValues(rerun, param_infl)
}
\arguments{
\item{rerun}{A list containing the new regression estimate, standard error covariance (\code{se_mat}), and standard errors (\code{se}).  The standard errors are the square root of the diagonal of \code{se_mat}, and can be interpreted as an estimate of the standard deviation of the coefficient estimates.}

\item{param_infl}{A parameter influence object (e.g. as produced by \link{AppendTargetRegressorInfluence}).}
}
\value{
A named numeric vector of base values.
}
\description{
Compute the "base values" (beta and the CI) for a parameter from a rerun.
}
