% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_lib.R
\name{GetSignalsAndRerunsDataframe}
\alias{GetSignalsAndRerunsDataframe}
\title{For signals and reruns lists, produce a dataframe summarizing the two.}
\usage{
GetSignalsAndRerunsDataframe(signals, reruns, model_grads)
}
\description{
For signals and reruns lists, produce a dataframe summarizing the two.
}
