% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_lib.R
\name{GetSortedInfluenceDf}
\alias{GetSortedInfluenceDf}
\title{Produce an influence dataframe suitable for visualization.}
\usage{
GetSortedInfluenceDf(param_infl, sorting_qoi_name, max_num_obs = Inf)
}
\arguments{
\item{param_infl}{A parameter influence object (e.g. as produced by \link{AppendTargetRegressorInfluence}).}

\item{sorting_qoi_name}{The name of a QOI ("beta", "beta_mzse", or
"beta_pzse") whose influence is used to sort the dataframe.}

\item{max_num_obs}{(Optional)  Include at most the \code{max_num_obs}
most influential observations for the sorting QOI.
Default is to include all observations.}
}
\value{
A dataframe with predictions, leaving out cumulatively more
points according to the sorting QOI's influence scores.
}
\description{
Produce an influence dataframe suitable for visualization.
}
