% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_iv_grads_lib.R
\name{GetTargetRegressorGrads}
\alias{GetTargetRegressorGrads}
\title{Get the influence scores for a particular regressor.}
\usage{
GetTargetRegressorGrads(reg_infl, target_regressor, sig_num_ses = qnorm(0.975))
}
\arguments{
\item{reg_infl}{The gradients, e.g. as computed by \code{\link[=ComputeModelInfluence]{ComputeModelInfluence()}}.}

\item{target_regressor}{The name of the regressor to target.  This should
correspond to the name of a column of the \code{x} matrix of the original fit.}

\item{sig_num_ses}{How wide the confidence interval is, as a number of standard errors.}
}
\value{
A gradient dataframe for the specified coefficient with extra
attributes set, suitable for passing to \code{\link[=SortAndAccumulate]{SortAndAccumulate()}}.
}
\description{
Get the influence scores for a particular regressor.
}
