% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_lib.R
\name{GetWeightForAlpha}
\alias{GetWeightForAlpha}
\title{Get a set of rows targeting a particular value of alpha.}
\usage{
GetWeightForAlpha(
  infl_df,
  alpha_colname,
  alpha_val,
  boolean = TRUE,
  rows_to_keep = TRUE
)
}
\arguments{
\item{infl_df}{A single influence dataframe.}

\item{alpha_colname}{A string with the name of the target alpha column}

\item{alpha_val}{The alpha value to target}

\item{boolean}{Optional.  If TRUE, return a boolean vector.  If FALSE,
return a vector of integer indices.}

\item{rows_to_keep}{Optional.  If TRUE, return a vector of rows to keep
for the target level of alpha.  If FALSE, return a vector of rows to drop.}
}
\value{
A row vector for the original dataframe corresponding to the target
level of alpha.  Note that the number of left-out rows will be rounded up.
}
\description{
Get a set of rows targeting a particular value of alpha.
}
