% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_lib.R
\name{GetWeightForTargetChangeRow}
\alias{GetWeightForTargetChangeRow}
\title{Get the weight vector to achieve the change given by a single row of
a target change.}
\usage{
GetWeightForTargetChangeRow(
  influence_dfs,
  target_change,
  boolean = TRUE,
  rows_to_keep = TRUE
)
}
\arguments{
\item{influence_dfs}{A list of influence dataframes}

\item{target_change}{A single row of a target change dataframe.}

\item{boolean}{Optional.  If TRUE, return a boolean vector.  If FALSE,
return a vector of integer indices.}

\item{rows_to_keep}{Optional.  If TRUE, return a vector of rows to keep
for the target level of alpha.  If FALSE, return a vector of rows to drop.}
}
\description{
Get the weight vector to achieve the change given by a single row of
a target change.
}
