% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_lib.R
\name{GetWeightVector}
\alias{GetWeightVector}
\title{Compute a weight vector for a set of dropped indices.}
\usage{
GetWeightVector(
  drop_inds,
  num_obs = NULL,
  orig_weights = NULL,
  bool = FALSE,
  invert = FALSE
)
}
\arguments{
\item{drop_inds}{\verb{The indices to drop (in the order of the original data)}}

\item{num_obs}{The number of observations in the original data.}

\item{bool}{(Optional)  If true, return a boolean vector.  Otherwise,
return a numeric vector (with ones and zeros).}

\item{invert}{(Optional) If TRUE, return a vector that retains the
observations in \code{drop_inds}.  Default is FALSE.}
}
\value{
A vector of weights in the order of the original data.
}
\description{
Compute a weight vector for a set of dropped indices.
}
