% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_lib.R
\name{InitializePython}
\alias{InitializePython}
\title{Start a Python session using the given Python 3 binary.}
\usage{
InitializePython(python_path = "/usr/bin/python3")
}
\arguments{
\item{python_path}{The path to a Python 3 binary.}
}
\value{
The \code{reticulate::py_main} object.
}
\description{
Start a Python session using the given Python 3 binary.
}
