% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_lib.R
\name{PairInfluenceScores}
\alias{PairInfluenceScores}
\title{Process influence dataframes to require observations to be removed in pairs.}
\usage{
PairInfluenceScores(
  influence_df,
  assignment_col,
  influence_cols = NULL,
  group_cols = NULL,
  level0 = 0,
  level1 = 1
)
}
\arguments{
\item{influence_df}{The gradient dataframe with attributes for a particular coefficient, e.g., as returned by \code{\link[=GetTargetRegressorGrads]{GetTargetRegressorGrads()}}.}

\item{assignment_col}{A string containing the column name of the assignment.}

\item{influence_cols}{Optional custom influence column names}

\item{group_cols}{Optional.  If you wish to make sure that additional
attributes match in removed pairs, you can specify the columns to match on
by passing the column names in as group_cols.}

\item{level0}{The first value of assignment.  Each pair will have one
with this assignment level.}

\item{level1}{The second value of assignment.  Each pair will also have one
with this assignment level.}
}
\value{
A list of influence dataframes analogous to that returned by
\code{\link[=SortAndAccumulate]{SortAndAccumulate()}}, but with paired, grouped observations.
}
\description{
It may be desireable to enforce removal of pairs of observations, e.g.,
to reomve the same number of control as treatment observations.  To
facilitate such analyses, this function pairs observations together into
"combined" observations which can then be analyzed with the rest of the
\code{zaminfluence} ecosystem.  Optionally, one may also enforce that pairs
match one another according to some group attributes.  The output of this
function can be used in place of the output of \code{\link[=SortAndAccumulate]{SortAndAccumulate()}}.
}
