% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_lib.R
\name{plot_influence_df}
\alias{plot_influence_df}
\title{Plot influence scores, signals, and reruns.}
\usage{
plot_influence_df(
  influence_df,
  signal,
  rerun_vals = NULL,
  plot_num_dropped = FALSE,
  apip_max = NULL,
  include_y_zero = TRUE
)
}
\arguments{
\item{influence_df}{The output of \link{get_sorted_influence_df}}

\item{signal}{A signal object (e.g. one of the elements of the list produced by \link{get_inference_signals_for_parameter}).}

\item{rerun_vals}{(Optional) Named list with entries \code{param}, \code{param_mzse},
\code{param_pzse} giving refit values to overlay on the plot.}

\item{plot_num_dropped}{If TRUE, plot the number dropped on the x-axis.
If FALSE (the default), plot the proportion dropped.}

\item{apip_max}{The maximum value for the x-axis (as a number or proportion
according to the value of \code{plot_num_dropped}).}

\item{include_y_zero}{(Optional) If TRUE (the default), force the y-axis
to include zero and plot a horizontal line.}
}
\value{
A plot.
}
\description{
Plot influence scores, signals, and reruns.
}
