% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_lib.R
\name{plot_signal}
\alias{plot_signal}
\title{Plot influence scores, signals, and reruns.}
\usage{
plot_signal(
  model_grads,
  signals,
  parameter_name,
  target_signal,
  reruns = NULL,
  ...
)
}
\arguments{
\item{param_infl}{A parameter influence object (e.g. as produced by \link{append_target_regressor_influence}).}

\item{signal}{A signal object (e.g. one of the elements of the list produced by \link{get_inference_signals_for_parameter}).}
}
\value{
A plot for the specified signal.
}
\description{
Plot influence scores, signals, and reruns.
}
