% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_lib.R
\name{PlotInfluence}
\alias{PlotInfluence}
\title{Produce a plot of the estimated effects of datapoint removal.}
\usage{
PlotInfluence(
  influence_dfs,
  alpha_colname,
  alpha_max,
  target_change = NULL,
  include_y_zero = TRUE,
  sig_num_ses = qnorm(0.975),
  rerun_df = NULL
)
}
\arguments{
\item{influence_dfs}{A list of sorted influence dataframes as returned, e.g., by \code{\link[=SortAndAccumulate]{SortAndAccumulate()}}.  The list is expected to have results sorted for both negative and positive changes in both sign and signifiance.}

\item{alpha_colname}{A string with the name of the target alpha column (typically \code{prop_removed} or \code{num_removed})}

\item{alpha_max}{The largest \code{alpha} value plotted, i.e., the rightmost
extent of the x-axis.}

\item{target_change}{A dataframe of target changes, e.g., as returned by \code{\link[=GetRegressionTargetChange]{GetRegressionTargetChange()}}}

\item{include_y_zero}{Whether to force the y-axis to include 0.}

\item{sig_num_ses}{How wide the confidence interval is, as a number of standard errors.}

\item{rerun_df}{Optional.  If specified, plot the results of
a rerun as produced by \code{RerunRegression}.}
}
\value{
A ggplot object containing a graphical summary of the estimate
influence.
}
\description{
Produce a plot of the estimated effects of datapoint removal.
}
