% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_lib.R
\name{predict_for_signals}
\alias{predict_for_signals}
\title{Predict the model at the AMIS for a set of signals.}
\usage{
predict_for_signals(signals, model_grads, verbose = FALSE)
}
\arguments{
\item{signals}{A list of signal objects, "sign", "sig", "both",
as produced by \link{get_inference_signals_for_parameter}.}

\item{model_grads}{A model gradient object (e.g. as produced by \link{compute_model_influence})}

\item{verbose}{(Optional) If TRUE, print status updates.}
}
\value{
A nested list of predictions matching the structure of \code{signals}.
}
\description{
Predict the model at the AMIS for a set of signals.
}
