% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_lib.R
\name{PredictChange}
\alias{PredictChange}
\title{Predict the effect of dropping points.}
\usage{
PredictChange(qoi, drop_inds)
}
\arguments{
\item{qoi}{\verb{A quantity of interest object (e.g. as produced by [QOIInfluence]).}}

\item{drop_inds}{\verb{The indices to drop (in the order of the original data)}}
}
\value{
A linear approximation to the effect of dropping the specified
observations.
}
\description{
Predict the effect of dropping points.
}
