% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_lib.R
\name{PredictForSignals}
\alias{PredictForSignals}
\title{Predict the model at the AMIS for a set of signals.}
\usage{
PredictForSignals(signals, model_grads, verbose = FALSE)
}
\arguments{
\item{signals}{`A list of signal objects, "sign", "sig", "both",
as produced by \link{GetInferenceSignalsForParameter}.}

\item{model_grads}{A model gradient object (e.g. as produced by \link{ComputeModelInfluence})}

\item{verbose}{(Optional) If TRUE, print status updates.}
}
\value{
The original \code{signals}, with two new fields:
\itemize{
\item rerun_df: A dataframe summarizing the rerun.
\item rerun: The complete rerun result.
}
}
\description{
Predict the model at the AMIS for a set of signals.
}
