% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_lib.R
\name{PrependZeroRemoved}
\alias{PrependZeroRemoved}
\title{Prepend a row for removing no rows and effecting no change to a sorted
gradient dataframe.}
\usage{
PrependZeroRemoved(df, keep_cols)
}
\description{
Prepend a row for removing no rows and effecting no change to a sorted
gradient dataframe.
}
