% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_lib.R
\name{QOIInfluence}
\alias{QOIInfluence}
\title{Process a vector of influence scores to produce sorted influence scores.}
\usage{
QOIInfluence(infl, base_value, num_obs = NULL)
}
\arguments{
\item{infl}{A vector of influence scores for a quantity of interest,
in the same order as the original data.}

\item{base_value}{The value of the quantity of interest at the original fit.}
}
\value{
See "Quantity of Interest" in README.md
}
\description{
Process a vector of influence scores to produce sorted influence scores.
}
