% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_lib.R
\name{RegressWithOffset}
\alias{RegressWithOffset}
\title{Compute a regression using the moment condition Cov(x, resid) = offset.}
\usage{
RegressWithOffset(lm_result, offset, se_group = NULL)
}
\arguments{
\item{lm_result}{The regression result, i.e, the output of \code{\link[=lm]{lm()}}.}

\item{offset}{The target moment condition.}

\item{se_group}{Optional. The standard error grouping variable.}
}
\value{
A list containing the regression and influence result.
}
\description{
Compute a regression using the moment condition Cov(x, resid) = offset.
}
