% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_lib.R
\name{rerun_for_signals}
\alias{rerun_for_signals}
\title{Rerun the model at the AMIS for a set of signals.}
\usage{
rerun_for_signals(signals, model_grads, rerun_fun = NULL, verbose = FALSE)
}
\arguments{
\item{signals}{A list of signal objects, "sign", "sig", "both",
as produced by \link{get_inference_signals_for_parameter}.}

\item{model_grads}{A model gradient object (e.g. as produced by \link{compute_model_influence})}

\item{rerun_fun}{(Optional) A function taking as inputs
a vector of weights, and returning a model_fit object.  If unspecified,
\code{model_grads$rerun_fun} is used.}

\item{verbose}{(Optional) If TRUE, print status updates.}
}
\value{
A nested list of reruns matching the structure of \code{signals}.
}
\description{
Rerun the model at the AMIS for a set of signals.
}
