% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_lib.R
\name{RerunForSignals}
\alias{RerunForSignals}
\title{Rerun the model at the AMIS for a set of signals.}
\usage{
RerunForSignals(signals, model_grads, RerunFun = NULL, verbose = FALSE)
}
\arguments{
\item{signals}{`A list of signal objects, "sign", "sig", "both",
as produced by \link{GetInferenceSignalsForParameter}.}

\item{model_grads}{A model gradient object (e.g. as produced by \link{ComputeModelInfluence})}

\item{RerunFun}{(Optional) A function taking as inputs
a vector of weights, and returning a ModelFit object.  If unspecified,
\code{model_grads$RerunFun} is used.}

\item{verbose}{(Optional) If TRUE, print status updates.}
}
\value{
The original \code{signals}, with two new fields:
\itemize{
\item rerun_df: A dataframe summarizing the rerun.
\item rerun: The complete rerun result.
}
}
\description{
Rerun the model at the AMIS for a set of signals.
}
