% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_targets_lib.R
\name{RerunForTargetChanges}
\alias{RerunForTargetChanges}
\title{Rerun the model at the AMIS for a set of signals.}
\usage{
RerunForTargetChanges(signals, model_grads, RerunFun)
}
\arguments{
\item{signals}{`A list of signal objects, "sign", "sig", "both",
as produced by \link{GetInferenceSignals}.}

\item{model_grads}{A model gradient object (e.g. as produced by \link{ComputeModelInfluence})}

\item{RerunFun}{A function taking as inputs \code{model_grads$model_fit}
and a vector of weights, and return a rerun object.}
}
\value{
The original \code{signals}, with two new fields:
\itemize{
\item rerun_df: A dataframe summarizing the rerun.
\item rerun: The complete rerun result.
}
}
\description{
Rerun the model at the AMIS for a set of signals.
}
