% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_iv_grads_lib.R
\name{RerunRegression}
\alias{RerunRegression}
\title{Rerun the regression with a new subset of rows.}
\usage{
RerunRegression(w_bool, lm_result, se_group = NULL, save_w = FALSE)
}
\arguments{
\item{w_bool}{A boolean vector of rows to keep in the original dataframe.}

\item{lm_result}{The regression result, i.e, the output of \code{\link[=lm]{lm()}}.}

\item{se_group}{Optional. The standard error grouping variable.}

\item{save_w}{Optional. If \code{TRUE}, save the new weight vector in the output.}
}
\value{
A list containing the new regression estimate, standard error covariance (\code{se_mat}), and standard errors (\code{se}).  The standard errors are the square root of the diagonal of \code{se_mat}, and can be interpreted as an estimate of the standard deviation of the coefficient estimates.
}
\description{
Rerun the regression with a new subset of rows.
}
