% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_lib.R
\name{RerunTargetModelForAlpha}
\alias{RerunTargetModelForAlpha}
\title{Rerun the target regression for a particular target alpha.}
\usage{
RerunTargetModelForAlpha(
  infl_df,
  model_fit,
  alpha_colname,
  alpha_val,
  se_group = NULL
)
}
\arguments{
\item{infl_df}{A particular influence dataframe.}

\item{model_fit}{The fit from `lm` or `AER:ivreg`.}

\item{alpha_colname}{A string with the name of the target alpha column}

\item{alpha_val}{The alpha value to target}

\item{se_group}{Optional. The standard error grouping variable.}
}
\value{
A dataframe with variables from the re-run model.
}
\description{
Rerun the target regression for a particular target alpha.
}
