% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_lib.R
\name{RerunTargetRegressorForAlpha}
\alias{RerunTargetRegressorForAlpha}
\title{Rerun the target regression for a particular target alpha.}
\usage{
RerunTargetRegressorForAlpha(
  infl_df,
  lm_result,
  alpha_colname,
  alpha_val,
  se_group = NULL
)
}
\arguments{
\item{infl_df}{A particular influence dataframe.}

\item{lm_result}{The original regression result.}

\item{alpha_colname}{A string with the name of the target alpha column}

\item{alpha_val}{The alpha value to target}

\item{se_group}{Optional. The standard error grouping variable.}
}
\value{
A dataframe with variables from the re-run regression.
}
\description{
Rerun the target regression for a particular target alpha.
}
