% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_lib.R
\name{SafeGetBaseVals}
\alias{SafeGetBaseVals}
\title{Return the $Gamma_alpha$ quantity.}
\usage{
SafeGetBaseVals(influence_dfs)
}
\arguments{
\item{influence_dfs}{A list of sorted influence dataframes as returned, e.g., by \code{\link[=SortAndAccumulate]{SortAndAccumulate()}}.  The list is expected to have results sorted for both negative and positive changes in both sign and signifiance.}
}
\value{
The <base_vals> attribute, having checked that both pos and neg
have the same values.
}
\description{
Return the $Gamma_alpha$ quantity.
}
