% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_lib.R
\name{SetPythonRegressionVariables}
\alias{SetPythonRegressionVariables}
\title{Set python variables corresponding to the output of lm.}
\usage{
SetPythonRegressionVariables(lm_res, se_group = NULL)
}
\description{
Set python variables corresponding to the output of lm.
}
