% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_lib.R
\name{SortAndAccumulate}
\alias{SortAndAccumulate}
\title{Sort and process a gradient dataframe to approximate adversarial removal sets.}
\usage{
SortAndAccumulate(grad_df, influence_cols = NULL)
}
\arguments{
\item{grad_df}{The output of GetTargetRegressorGrads.}
}
\value{
A list of lists, ultimately containing dataframes.  The entries
sign and sig refer to changing sign and significance of the target regressor,
and entries pos and neg refer to increasing or decreasing.  See
SortAndAccumulateForSign for details of the dataframe.
}
\description{
Sort and process a gradient dataframe to approximate adversarial removal sets.
}
