% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_lib.R
\name{SortInfluenceDf}
\alias{SortInfluenceDf}
\title{Sort a gradient dataframe to effect a change in the direction of
<change_sign> using <influence_col>.  Append columns for the number and
proportion of rows removed.}
\usage{
SortInfluenceDf(influence_df, influence_col, change_sign)
}
\description{
Sort a gradient dataframe to effect a change in the direction of
<change_sign> using <influence_col>.  Append columns for the number and
proportion of rows removed.
}
