% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_iv_grads_torch_lib.R
\name{compute_model_influence}
\alias{compute_model_influence}
\title{Compute the influence functions for all regressors given a model fit.}
\usage{
compute_model_influence(fit_object, se_group = NULL, keep_pars = NULL)
}
\arguments{
\item{fit_object}{The fit from \code{\link[=lm]{lm()}} or \code{\link[ivreg:ivreg]{ivreg::ivreg()}}.}

\item{se_group}{Optional. The standard error grouping variable.}

\item{keep_pars}{Optional. The names of parameters for which to compute derivatives.}
}
\value{
A model gradient object (e.g. as produced by \link{compute_model_influence})
}
\description{
Compute the influence functions for all regressors given a model fit.
}
