% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_iv_grads_lib.R
\name{ComputeModelInfluence}
\alias{ComputeModelInfluence}
\title{Compute the influence functions for all regressors given a model fit.}
\usage{
ComputeModelInfluence(model_fit, se_group = NULL)
}
\arguments{
\item{model_fit}{The fit from \code{\link[=lm]{lm()}} or \code{\link[AER:ivreg]{AER::ivreg()}}.}

\item{se_group}{Optional. The standard error grouping variable.}
}
\value{
A list containing the sensitivity of the coefficients and standard errors to data re-weighting.  The list entries are
\describe{
\item{model_fit}{The original fit as returned by the fitting function}
\item{n_obs}{The number of observations given weights}
\item{regressor_names}{The names of the regressors as strings}
\item{grad_fun}{The function used to compute the derivatives}
\item{betahat}{The estimated regression coefficient at the original weights}
\item{se}{The standard errors at the original weights}
\item{weights}{The original weights}
\item{beta_grad}{The gradient of betahat with respect to the weights}
\item{se_grad}{The gradient of se with respect to the weights}
}
}
\description{
Compute the influence functions for all regressors given a model fit.
}
