% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_iv_grads_lib.R
\name{ComputeModelInfluence}
\alias{ComputeModelInfluence}
\title{Compute the influence functions for all regressors given a model fit.}
\usage{
ComputeModelInfluence(model_fit, se_group = NULL)
}
\arguments{
\item{model_fit}{A model fit (currently from lm or AER::iv_reg).}

\item{se_group}{Optional, a vector of integers defining a standard error grouping.}
}
\value{
A list containing the regression and influence result.
}
\description{
Compute the influence functions for all regressors given a model fit.
}
