% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_lib.R
\name{ComputeRegressionMomentSensitivity}
\alias{ComputeRegressionMomentSensitivity}
\title{Compute the sensitivity of a regression to the moment condition on the
residuals.  That is, compute the gradients with respect to \code{offset},
where the regression is fit with the moment condition Cov(x, resid) = offset.}
\usage{
ComputeRegressionMomentSensitivity(lm_result, se_group = NULL)
}
\arguments{
\item{lm_result}{The regression result, i.e, the output of \code{\link[=lm]{lm()}}.}

\item{se_group}{Optional. The standard error grouping variable.}
}
\value{
A list containing the regression and influence result.
}
\description{
Compute the sensitivity of a regression to the moment condition on the
residuals.  That is, compute the gradients with respect to \code{offset},
where the regression is fit with the moment condition Cov(x, resid) = offset.
}
