% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_iv_grads_lib.R, R/ols_iv_grads_torch_lib.R
\name{ComputeRegressionResults}
\alias{ComputeRegressionResults}
\title{Run a regression using zaminfluence code.  This should be identical
to ordinary regression.}
\usage{
ComputeRegressionResults(lm_result, weights = NULL, se_group = NULL)

ComputeRegressionResults(lm_result, weights = NULL, se_group = NULL)
}
\arguments{
\item{lm_result}{The regression result, i.e, the output of \code{\link[=lm]{lm()}}.}

\item{weights}{Optional.  A vector of weights.  If unset, use the original weights.}

\item{se_group}{Optional. The standard error grouping variable.}
}
\value{
A list containing the new regression estimate, standard error covariance (\code{se_mat}), and standard errors (\code{se}).  The standard errors are the square root of the diagonal of \code{se_mat}, and can be interpreted as an estimate of the standard deviation of the coefficient estimates.

A list containing the new regression estimate, standard error covariance (\code{se_mat}), and standard errors (\code{se}).  The standard errors are the square root of the diagonal of \code{se_mat}, and can be interpreted as an estimate of the standard deviation of the coefficient estimates.
}
\description{
Run a regression using zaminfluence code.  This should be identical
to ordinary regression.

Run a regression using zaminfluence code.  This should be identical
to ordinary regression.
}
