% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_lib.R
\name{get_signals_and_reruns_df}
\alias{get_signals_and_reruns_df}
\title{For signals and reruns lists, produce a dataframe summarizing the two.}
\usage{
get_signals_and_reruns_df(signals, reruns, model_grads)
}
\description{
For signals and reruns lists, produce a dataframe summarizing the two.
}
