% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_lib.R
\name{GetAMIP}
\alias{GetAMIP}
\title{Compute the approximate maximum influence perturbation (AMIP).}
\usage{
GetAMIP(qoi, sign, n_drop)
}
\arguments{
\item{qoi}{\verb{A quantity of interest object (e.g. as produced by [QOIInfluence]).}}

\item{n_drop}{The number of points to drop (we will round up).}
}
\value{
The approximate largest change that can be produced by dropping
the specified number of points.
}
\description{
Compute the approximate maximum influence perturbation (AMIP).
}
