% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_lib.R
\name{GetAMIS}
\alias{GetAMIS}
\title{Compute the approximate maximally-influential set (AMIS).}
\usage{
GetAMIS(qoi, n_drop = NULL, prop_drop = NULL)
}
\arguments{
\item{qoi}{\verb{A quantity of interest object (e.g. as produced by [ProcessInfluenceVector]).}}

\item{n_drop}{The number of points to drop (we will round up).
Exactly one of \code{n_drop} and \code{prop_drop} must be specified.}

\item{prop_drop}{The proportion of points to drop (we will round up).
Exactly one of \code{n_drop} and \code{prop_drop} must be specified.}
}
\value{
The indices to drop (in the order of the original data)
}
\description{
Compute the approximate maximally-influential set (AMIS).
}
