% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_targets_lib.R
\name{PlotInfluence}
\alias{PlotInfluence}
\title{Plot influence scores, signals, and reruns.}
\usage{
PlotInfluence(
  influence_df,
  plot_num_dropped = FALSE,
  apip_max = NULL,
  signals = NULL,
  include_y_zero = TRUE
)
}
\arguments{
\item{influence_df}{The output of \link{GetSortedInfluenceDf}}

\item{plot_num_dropped}{If TRUE, plot the number dropped on the x-axis.
If FALSE (the default), plot the proportion dropped.}

\item{apip_max}{The maximum value for the x-axis (as a number or proportion
according to the value of \code{plot_num_dropped}).}

\item{signals}{(Optional) A list of signals to plot.}

\item{include_y_zero}{(Optional) If TRUE (the default), force the y-axis
to include zero and plot a horizontal line.}
}
\value{
A plot.
}
\description{
Plot influence scores, signals, and reruns.
}
