% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_targets_lib.R
\name{PlotSignal}
\alias{PlotSignal}
\title{Plot influence scores, signals, and reruns.}
\usage{
PlotSignal(param_infl, signal, ...)
}
\arguments{
\item{param_infl}{A parameter influence object (e.g. as produced by \link{AppendTargetRegressorInfluence}).}

\item{signal}{A signal object (e.g. one of the elements of the list produced by \link{GetInferenceSignalsForParameter}).}
}
\value{
A plot for the specified signal.
}
\description{
Plot influence scores, signals, and reruns.
}
