% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_lib.R
\name{RerunForTargetChanges}
\alias{RerunForTargetChanges}
\title{Rerun the target regression for every change in target_change.}
\usage{
RerunForTargetChanges(influence_dfs, target_change, model_fit, se_group = NULL)
}
\arguments{
\item{influence_dfs}{A list of influence dataframes}

\item{target_change}{The output of GetRegressionTargetChange}

\item{model_fit}{The fit from \code{lm} or \code{AER:ivreg}.}
}
\value{
A dataframe with variables from all the re-run models.
}
\description{
Rerun the target regression for every change in target_change.
}
