% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_lib.R
\name{RerunForTargetChanges}
\alias{RerunForTargetChanges}
\title{Rerun the target regression for every change in target_change.}
\usage{
RerunForTargetChanges(influence_dfs, target_change, model_fit, se_group = NULL)
}
\arguments{
\item{influence_dfs}{Influence dataframes, dude!}

\item{target_change}{A single row of the target change dataframe produced
by \code{\link[=GetRegressionTargetChange]{GetRegressionTargetChange()}}.}

\item{model_fit}{The fit from \code{\link[=lm]{lm()}} or \code{\link[AER:ivreg]{AER::ivreg()}}.}
}
\value{
A list containing the new regression estimate, standard error covariance (\code{se_mat}), and standard errors (\code{se}).  The standard errors are the square root of the diagonal of \code{se_mat}, and can be interpreted as an estimate of the standard deviation of the coefficient estimates.
}
\description{
Rerun the target regression for every change in target_change.
}
