% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_lib.R
\name{RerunIVRegression}
\alias{RerunIVRegression}
\title{Rerun the regression with a new subset of rows.}
\usage{
RerunIVRegression(w_bool, iv_res, se_group = NULL, save_w = FALSE)
}
\arguments{
\item{w_bool}{A boolean vector of rows to keep in the original dataframe.}

\item{iv_res}{The iv regression result, i.e, the output of \code{\link[=ivreg]{ivreg()}}.}

\item{se_group}{Optional. The standard error grouping variable.}

\item{save_w}{Optional. If TRUE, save the new weight vector in the output.}
}
\value{
A list containing the new regression estimate, standard error covariance (\code{se_mat}), and standard errors (\code{se}).  The standard errors are the square root of the diagonal of \code{se_mat}, and can be interpreted as an estimate of the standard deviation of the coefficient estimates.
}
\description{
Rerun the regression with a new subset of rows.
}
