% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_lib.R
\name{RerunRegression}
\alias{RerunRegression}
\title{Rerun the regression with a new subset of rows.}
\usage{
RerunRegression(w_bool, lm_result, se_group = NULL, save_w = FALSE)
}
\arguments{
\item{w_bool}{A boolean vector of rows to keep in the original dataframe.}

\item{lm_result}{The original regression result.}

\item{se_group}{Optional. The standard error grouping variable.}

\item{save_w}{Optional. If TRUE, save the new weight vector in the output.}
}
\value{
A list containing the new regression estimate, standard error
covariance, and standard errors.
}
\description{
Rerun the regression with a new subset of rows.
}
