% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_lib.R
\name{RerunTargetIVRegressionForAlpha}
\alias{RerunTargetIVRegressionForAlpha}
\title{Rerun the target regression for a particular target alpha.}
\usage{
RerunTargetIVRegressionForAlpha(
  infl_df,
  iv_res,
  alpha_colname,
  alpha_val,
  se_group = NULL
)
}
\arguments{
\item{infl_df}{A single influence dataframe, dude!}

\item{iv_res}{The iv regression result, i.e, the output of \code{\link[=ivreg]{ivreg()}}.}

\item{alpha_colname}{A string with the name of the target alpha column (typically \code{prop_removed} or \code{num_removed})}

\item{alpha_val}{The alpha value to target}

\item{se_group}{Optional. The standard error grouping variable.}
}
\value{
A list containing the new regression estimate, standard error covariance (\code{se_mat}), and standard errors (\code{se}).  The standard errors are the square root of the diagonal of \code{se_mat}, and can be interpreted as an estimate of the standard deviation of the coefficient estimates.
}
\description{
Rerun the target regression for a particular target alpha.
}
