% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_lib.R
\name{SafeGetBaseVals}
\alias{SafeGetBaseVals}
\title{Return the $Gamma_alpha$ quantity.}
\usage{
SafeGetBaseVals(influence_dfs)
}
\arguments{
\item{influence_dfs}{A list with entries pos and neg, each of which is a
gradient dataframe.}
}
\value{
The <base_vals> attribute, having checked that both pos and neg
have the same values.
}
\description{
Return the $Gamma_alpha$ quantity.
}
