% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_lib.R
\name{SetPythonRegressionVariables}
\alias{SetPythonRegressionVariables}
\title{Deprecated --- use the functions in \code{ols_iv_grads_lib.R} instead.
Set python variables corresponding to the output of lm.}
\usage{
SetPythonRegressionVariables(lm_res, se_group = NULL)
}
\description{
Deprecated --- use the functions in \code{ols_iv_grads_lib.R} instead.
Set python variables corresponding to the output of lm.
}
