% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_lib.R
\name{SortAndAccumulateForSign}
\alias{SortAndAccumulateForSign}
\title{Sort and aggregate a gradient dataframe for a particular kind of change.}
\usage{
SortAndAccumulateForSign(grad_df, sort_col, change_sign, influence_cols)
}
\arguments{
\item{grad_df}{The output of GetTargetRegressorGrads.}

\item{sort_col}{Which column to sort by, typically a weight influence column.}

\item{change_sign}{The sign change that is to be effected by removing rows.}

\item{influence_cols}{Additional influence columns to be accumulated after
sorting, possibly including sort_col.}
}
\value{
A dataframe sorted using sort_col to effect a change in the
direction change_sign.  The following columns are calculated for each
column in influence_cols:
num_removed: The number of rows removed by removing this row and all above
prop_removed: The proportaion of rows removed by removing this row and all
above<col>_grad: The gradient of <col> for this row.
<col>_change: The estimated change in <col> when removing this row and all
previous rows (that is, the cumulative sum of the gradient).
<col>_est: The estimated value of <col> (that is, the original value plus
the estimated change).
}
\description{
Sort and aggregate a gradient dataframe for a particular kind of change.
}
